+++ android/core/designsystem/src/main/kotlin/com/andlife/designsystem/component/datepicker/state/DatePickerState.kt
@@ -1,6 +1,9 @@
 package com.andlife.designsystem.component.datepicker.state
 
+import androidx.compose.runtime.Composable
 import androidx.compose.runtime.Stable
+import androidx.compose.runtime.saveable.rememberSaveable
+import com.andlife.designsystem.component.datepicker.model.DatePickerDate
 import com.andlife.designsystem.component.datepicker.model.DatePickerYearMonth
 import com.andlife.designsystem.component.datepicker.ui.DatePickerMode
 import kotlinx.datetime.LocalDate
@@ -25,3 +28,15 @@ interface DatePickerState {
 
     fun showCalendar(yearMonth: DatePickerYearMonth)
 }
+
+@Composable
+fun rememberDatePickerState(
+    initialSelectedDate: LocalDate? = null,
+    initialMode: DatePickerMode = DatePickerMode.DATE,
+): DatePickerState =
+    rememberSaveable(saver = DatePickerStateImpl.Saver()) {
+        DatePickerStateImpl(
+            initialSelectedDate = initialSelectedDate?.let { DatePickerDate(it) },
+            initialMode = initialMode,
+        )
+    }
